django/requirements.txt
Django>=2.1.3,<3.3.0
django-cors-headers==3.10.0
djangorestframework>=3.12.4,<3.13.0
django-stubs>=1.9.0,<2.0.0
mypy>=0.910,<1.000
gunicorn==20.1.0
